Reorganize python modules
