+++ observatoire/tmdb/movies/helpers.py
@@ -1,22 +1,6 @@
-import json
 import logging
 
 import pandas as pd
-
-
-def parse_keywords(keywords: str) -> str:
-    """
-    Parses the keywords into a list
-    """
-
-    keywords_dict = json.loads(keywords)
-
-    keywords_list = []
-
-    for item in keywords_dict["keywords"]:
-        keywords_list.append(item["name"])
-
-    return ", ".join(keywords_list)
 
 
 def merge(df_current: pd.DataFrame | None, df_latest: pd.DataFrame, logger: logging) -> None:
